* Added ridge regression and pls-regression * Added base-class for linear-regression models * Added tests for new models
