+++ tests/models/test_model.py
@@ -4,7 +4,9 @@ import pytest
 
 from datamodels import (
     Model, 
-    LinearRegression, 
+    LinearRegression,
+    RidgeRegression,
+    PLSRegression,
     RandomForestRegression,
     SupportVectorRegression,
     XGBoost,
@@ -19,7 +21,9 @@ from datamodels import (
 
 
 @pytest.mark.parametrize('model_class', [
-    LinearRegression, 
+    LinearRegression,
+    RidgeRegression,
+    PLSRegression,
     RandomForestRegression,
     NeuralNetwork,
     VanillaLSTM,
@@ -62,7 +66,9 @@ def test_predict_sequence_save_and_load(model_class, tmpdir):
     assert y_pred.shape == (5, lookahead + 1, number_target_features)
 
 @pytest.mark.parametrize('model_class', [
-    LinearRegression, 
+    LinearRegression,
+    RidgeRegression,
+    PLSRegression,
     RandomForestRegression,
     SupportVectorRegression,
     XGBoost,
